Commit Implementado e validado as mensagens de retorno

+++ roadeye/src/main/java/buildrun/roadeye/rest/service/implementation/SchoolAddressServiceImplementation.java
@@ -1,21 +1,25 @@
-package buildrun.roadeye.rest.dto.service.implementation;
+package buildrun.roadeye.rest.service.implementation;
 
 import buildrun.roadeye.domain.entity.*;
 import buildrun.roadeye.domain.repository.AddressRepository;
 import buildrun.roadeye.domain.repository.SchoolAddressRepository;
 import buildrun.roadeye.domain.repository.SchoolRepository;
 import buildrun.roadeye.rest.dto.SchoolAddressDto;
-import buildrun.roadeye.rest.dto.service.SchoolAddressService;
+import buildrun.roadeye.rest.dto.UserAddressDto;
+import buildrun.roadeye.rest.service.SchoolAddressService;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 @Service
+@Transactional
 public class SchoolAddressServiceImplementation implements SchoolAddressService {
     private final SchoolAddressRepository schoolAddressRepository;
     private final SchoolRepository schoolRepository;
@@ -27,57 +31,61 @@ public class SchoolAddressServiceImplementation implements SchoolAddressService 
     }
 
     @Override
-    @Transactional
     public List<SchoolAddress> getAllSchoolAddress() {
         return schoolAddressRepository.findAll();
     }
 
     @Override
-    @Transactional
     public List<SchoolAddress> findAddressesBySchool_Id(Long schoolId) {
         return schoolAddressRepository.findBySchool_Id(schoolId);
     }
 
     @Override
-    @Transactional
-    public void deleteUserAddress(Long userAddressId) {
-        SchoolAddress schoolAddress = schoolAddressRepository.findById(userAddressId).orElseThrow(() -> new EntityNotFoundException("School or Address not found"));
-        schoolAddressRepository.delete(schoolAddress);
+    public ResponseEntity<?> deleteUserAddress(Long userAddressId) {
+        Optional<SchoolAddress> optionalSchoolAddress= schoolAddressRepository.findById(userAddressId);
+        if (optionalSchoolAddress.isPresent()) {
+            SchoolAddress schoolAddress = optionalSchoolAddress.get();
+            schoolAddressRepository.delete(schoolAddress);
+            ErrorResponse errorResponse = new ErrorResponse("School Address deleted successfully");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        } else {
+            ErrorResponse errorResponse = new ErrorResponse("School Address does not exist.");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        }
     }
 
     @Override
-    @Transactional
-    public SchoolAddress getUserAddressById(Long schoolAddressId) {
-        return schoolAddressRepository.findById(schoolAddressId).orElseThrow(() -> new EntityNotFoundException("SchoolAddress not found"));
-    }
-
-    @Override
-    @Transactional
-    public SchoolAddress updateSchoolAddress(Long schoolAddressId, SchoolAddressDto updateSchoolAddressDto) {
+    public ResponseEntity<?> updateSchoolAddress(Long schoolAddressId, SchoolAddressDto updateSchoolAddressDto) {
         Long schoolId = updateSchoolAddressDto.schoolId();
         Long addressId = updateSchoolAddressDto.addressId();
-
         Optional<School> optionalSchool = schoolRepository.findById(schoolId);
-        if (optionalSchool.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "School not found");
-        }
         Optional<Address> optionalAddress = addressRepository.findById(addressId);
-        if (optionalAddress.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Address not found");
-        }
-
         Optional<SchoolAddress> optionalSchoolAddress = schoolAddressRepository.findById(schoolAddressId);
         if (optionalSchoolAddress.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "UserAddress not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("SchoolAddress does not exist."));
+        }
+        if (optionalSchool.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("School does not exist."));
+        }
+        if (optionalAddress.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse("Address does not exist."));
         }
         School school = optionalSchool.get();
         Address address = optionalAddress.get();
-
         SchoolAddress schoolAddress = optionalSchoolAddress.get();
         schoolAddress.setSchool(school);
         schoolAddress.setAddress(address);
+        return ResponseEntity.ok(schoolAddressRepository.save(schoolAddress));
+    }
 
-        SchoolAddress savedAddress = schoolAddressRepository.save(schoolAddress);
-        return savedAddress;
+    @Override
+    public ResponseEntity<?> getSchoolAddressById(Long schoolAddressId) {
+        SchoolAddress schoolAddress = schoolAddressRepository.findById(schoolAddressId).orElse(null);
+        if (schoolAddress != null) {
+            return ResponseEntity.ok(schoolAddress);
+        } else {
+            ErrorResponse errorResponse = new ErrorResponse("School Address does not exist.");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        }
     }
 }
